+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added subset option to mean_dyns function.
 * Added vline option to plot.
 * Added common 31P brain metabolites.
+* Added verbose option to sim_basis.
 
 # spant 2.15.0
 * An SNR estimate of zero in comb_coils is now a failure case.